Refactor app into blueprints with Korean comments
