+++ templates/carriers.html
@@ -5,7 +5,7 @@
     Name <input type="text" name="carrier_name" value="{{ request.args.get('carrier_name','') }}">
     Type <input type="text" name="type" value="{{ request.args.get('type','') }}">
     <input type="submit" value="Search">
-    <a href="{{ url_for('new_carrier') }}">Create</a>
+    <a href="{{ url_for('carriers.new_carrier') }}">Create</a>
 </form>
 <table>
 <tr>
@@ -26,8 +26,8 @@
     <td>{{ c['supports_freezing'] }}</td>
     <td>{{ c['supports_dangerous_goods'] }}</td>
     <td>
-        <a href="{{ url_for('edit_carrier', id=c['id']) }}">Edit</a>
-        <form class="inline" method="post" action="{{ url_for('delete_carrier', id=c['id']) }}">
+        <a href="{{ url_for('carriers.edit_carrier', id=c['id']) }}">Edit</a>
+        <form class="inline" method="post" action="{{ url_for('carriers.delete_carrier', id=c['id']) }}">
             <button type="submit">Delete</button>
         </form>
     </td>
