+++ templates/locations.html
@@ -6,7 +6,7 @@
     Type <input type="text" name="type" value="{{ request.args.get('type','') }}">
     Country <input type="text" name="country" value="{{ request.args.get('country','') }}">
     <input type="submit" value="Search">
-    <a href="{{ url_for('new_location') }}">Create</a>
+    <a href="{{ url_for('locations.new_location') }}">Create</a>
 </form>
 <table>
 <tr>
@@ -22,8 +22,8 @@
     <td>{{ l['supports_storage'] }}</td>
     <td>{{ l['supports_dangerous_goods'] }}</td>
     <td>
-        <a href="{{ url_for('edit_location', id=l['id']) }}">Edit</a>
-        <form class="inline" method="post" action="{{ url_for('delete_location', id=l['id']) }}">
+        <a href="{{ url_for('locations.edit_location', id=l['id']) }}">Edit</a>
+        <form class="inline" method="post" action="{{ url_for('locations.delete_location', id=l['id']) }}">
             <button type="submit">Delete</button>
         </form>
     </td>
