+++ templates/routes.html
@@ -12,7 +12,7 @@
         {% endfor %}
     </select>
     <input type="submit" value="Search">
-    <a href="{{ url_for('new_route') }}">Create</a>
+    <a href="{{ url_for('routes.new_route') }}">Create</a>
 </form>
 <table>
 <tr>
@@ -31,8 +31,8 @@
     <td>{{ r['supports_freezing'] }}</td>
     <td>{{ r['supports_dangerous_goods'] }}</td>
     <td>
-        <a href="{{ url_for('edit_route', id=r['id']) }}">Edit</a>
-        <form class="inline" method="post" action="{{ url_for('delete_route', id=r['id']) }}">
+        <a href="{{ url_for('routes.edit_route', id=r['id']) }}">Edit</a>
+        <form class="inline" method="post" action="{{ url_for('routes.delete_route', id=r['id']) }}">
             <button type="submit">Delete</button>
         </form>
     </td>
